251031/Leontev Danil/project/project 3/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pydantic==2.6.1
python-dotenv==1.0.1
alembic==1.13.1
pytest==8.0.0
httpx==0.26.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
cassandra-driver==3.28.0
cassandra-cqlengine==3.1.1
